# openapi-test.yml
openapi: '3.0.3'
info:
  title: Praktikum REST API 2
  version: '1.0'
servers:
  - url: http://localhost:3000
components:
  schemas:
    user:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        age:
          type: string
        creatdAt:
          type: string
        updatedAt:
          type: string
paths:
  /user:
    get:
      description: 'Endpoint untuk Get all users'
      parameters: []
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/user'
        '500':
          description: Server Gagal
        '403':
          description: Tidak ada akses

    post:
      description: 'Endpoint untuk Add User'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user'
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '500':
          description: Server Gagal
        '403':
          description: Tidak ada akses

  /users/{id}:
    get:
      description: 'Endpoint untuk Get user by id'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '500':
          description: Server Gagal
        '403':
          description: Tidak ada akses
      
    delete:
      description: 'Endpoint untuk Delete user by id'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Berhasil
        '500':
          description: Server Gagal
        '403':
          description: Tidak ada akses
    put:
      description: 'Endpoint untuk Update user by id'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: email
          in: query
          required: true
          schema:
            type: string
        - name: age
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'